ingestion/requirements.txt
# Azure SDK
azure-identity>=1.15.0
azure-core>=1.30.0
azure-storage-blob>=12.19.0
azure-storage-file-datalake>=12.14.0
azure-search-documents>=11.6.0
azure-ai-documentintelligence>=1.0.0

# OpenAI
openai>=1.0.0

# Document parsing
pypdf>=4.0.0
pymupdf>=1.24.0
beautifulsoup4>=4.12.0
Pillow>=10.0.0

# Text processing
tiktoken>=0.5.0

# HTTP client
aiohttp>=3.9.0

# Retry logic
tenacity>=8.2.0

# Progress display
rich>=13.0.0

# Type hints
typing_extensions>=4.0.0
